+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+### Added
+- Add ability for partiql-extension-ion extension encoding/decoding of `Value` to/from Ion `Element`
+### Fixes
 
 ## [0.5.0] - 2023-06-06
 ### Changed